restructure and make OceanEnsembles.jl package
