+++ src/OceanEnsembles.jl
@@ -1,0 +1,9 @@
+module OceanEnsembles
+
+export basin_mask, get_coords_from_grid
+
+include("BasinMask.jl")
+
+using .BasinMask
+
+end # module OceanEnsembles
