+++ test/runtests.jl
@@ -1,0 +1,5 @@
+using OceanEnsembles, Test
+
+a = 1
+
+@test a == 1
